Add files via upload
Paper on implemented project titled: " Controlling Unsupervised Implantable  Robotic Arm Using Brain Computing Interface Technology"
Brain Computing Interface(BCI) is  technology for reading the brainwaves (the signals generated due to firing of neurons in the brain). the waves can be interpreted for the activities/emotions/thoughts in the brain like meditation, attention, sleeping, etc., and also explicit muscle activity like intentional eye blink.
The project focuses on using the attention waves and intentional eye blink to trigger and the control the actions of an implantable robotic arm. 
The AtMega8 microcontroller and L293DNE dual motor driver are used as driving parts of the robotic arm.